swm: wft/prepare_package -- progress secondary packages once we have a version
Currently we wait until the primary package is uploaded before we liven
the secondary prepare-package tasks.  This is very delayed.  As the
first thing we do is to add a version use that as an indication that
we have started working on this package and liven all relevant
prepare-package tasks.

Signed-off-by: Andy Whitcroft <apw@canonical.com>