maintscripts: verify-release-ready -- replace manual package lists
We are maintaining package lists in order to determine the type of
a package.  This information is maintained in the kernel-series
information, and we are already consulting that to determine if
the series is valid.  Use the sources and package lists to find
the package in question and identify its type.

Signed-off-by: Andy Whitcroft <apw@canonical.com>