cranky: rebase helper
This rebase script is going to be copied to the tree and be called by
cranky-rebase. It can also be called directly. And options may be given to
override its configuration, allowing embargoed uploads to be prepared.

There is no dependency besides bash, git and gawk. Most importantly, no
libraries that need to be copied to the tree or imported from something
that is not in the archive.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Marcelo Cerri <marcelo.cerri@canonical.com>