cranky-fix: resync in-tree helper scripts
Copy helper scripts from kteam-tools. These helper scripts are intended to
allow the kernel repository to be updated, so a new kernel can be prepared.
The list of scripts to be resynced needs to be updated. This is on purpose,
so we don't add files that are not ready to be updated.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Marcelo Cerri <marcelo.cerri@canonical.com>