cranky-fix: update update.conf with master debian dir
Using kernel series to determine the branch one kernel derives from, use that
branch debian/debian.env to determine which is the master debian directory.

With a raspi2 kernel that derives from a master kernel, that results in
debian.master. For a xenial/azure kernel that derives from a bionic/azure
kernel, that results in debian.azure.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Marcelo Cerri <marcelo.cerri@canonical.com>