helpers: rebase: look only at the master debian directory
Some git history have incorrect messages, missing the package name suffix.
Instead of Ubuntu-raspi2-4.4.0-1033.40, we have Ubuntu-4.4.0-1033.40. This
breaks the rebase helper.

If we restrict the log to changes to the master debian directory, that
works fine. We use the DEBIAN_MASTER variable set on updates.conf.

If DEBIAN_MASTER is not defined, tell the user where to set it. If DEBIAN is
debian.master, there should be no need to rebase.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Marcelo Cerri <marcelo.cerri@canonical.com>