cranky: close helper
This script is going to be copied to the tree and be called by "cranky
close". It can also be called directly.

There is no dependency besides bash, git, set and dpkg-parsechangelog.
Most importantly, no libraries that need to be copied to the tree or
imported from something that is not in the archive.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Marcelo Cerri <marcelo.cerri@canonical.com>