sru-cycle-kernel-versions: Update print format
Now that there is bionic and having hwe kernels there as well,
the old format is becoming more confusing.
This re-arranges output to group kernels by series and also
does order the series in a latest to oldest sequence (from a
release point of view). The output format chosen would also
allow the info to be parsed as yaml.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>