swm: wft/promote_to_release -- if we are not in proposed or not tested keep quiet
We should not be checking for publication of the package to the release
pocket if it is not yet in proposed, or if its testing is not complete.
Short circuit our checks so that we do not produce unhelpful reason
messages when the sources are not uploaded.  Cargo cult checks from
promote_to_updates for this.

Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>