swm: manager -- handle status updates with concurrent runners
With locking granularity dropped to a bug level it is entirely possible
for us to have more than one bug being shanked at the same time.  We
need to mediate access to the status so we do not remove other runners
updates.

Add a new lock for the status, using a bug lock on bug #1.  Under this
lock, load, update and save the status atomically when updating a single
entry.  When doing a full run (which allows us to purge dead entries)
only consider an entry for deletion if it was present when we started.
Any other entries must be from parallel writers and we have no information
on their validility (it is likely valid as they are new); they will be
reaped in the next run in the unlikely event they are then no longer open.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>