swm: wft/promote_to_proposed -- handle removed source when copy is in progress
When we have successfully copied kernels but they are not completely
published (as happens if the bits get caught in Unapproved or New)
then we remove the original source, we will detect the loss of source
and throw the promote-to-proposed into Incomplere.  We should only
do that if the binaries are not published.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>