cranky-start: force maint-startnewrelease to be run locally
Starting with 4.18, we need to run cranky-start on the appropriate
schroot and that causes some issues when maint-startnewrelease tries to
reach the remote server via SSH and the SSH keys are not available on
the schroot.

Force cranky-start to alreasy run maint-startnewrelease locally. That
is a temporary workaround while cranky-start is not rewritten. Also
the impact of having to download the packages locally nowadays os very
small.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>