swm: run workflow task processing against the primary task
We are running work for the primary task within the promote-to-updates
and promote-to-release tasks.  Actually this is work about the state of
the primary tasks so it makes more sense to evaluate that tasks status
for itself.  Provide a handler for this task and let it assert the
closing criteria.  This allows us to report failure to close reasons
against the primary task.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>