swm: bug -- versions are valid when publication is pending
When publication is pending, ie when the publisher has been told
something is ready to publish but it is not yet actually published, we
actually have a completely valid version.  It is not published enough to
start testing but it is valid for version comparisons.  Allow pending
publication states to supply versions.

This is visible particularly for git-tags checks which will report
missing tags if we have no version available.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>