maintscripts: Stop treating tilde in versions special
For quite a while now maint-startnewrelease would not try to
get the ABI files if the version number had a '~' in it. This
was a hack that started when we introduced LTS backport kernels
which lazily skipped fetching any ABI files.

That abuse was also ported to maint-getabis and needs to be
stopped there as well.

While on it, improve the dry-run output a little and only
try to delete the previous ABI files if it has not already
been done by any nested script.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Kleber Souza <kleber.souza@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>