adt-matrix: ignore britney hints for primary packages
It is lovely to honour the britney hints for test packages, but this
leads us to potentially let out kernel packages based on those hints.
While that is valid for a secondary package it is concerning for primary
packages.  Elide hints for the primary package so we have to explicitly
hint those in our kernel hints.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>