adt-matrix: ignore britney test list before populated
It seems that britney includes the 'autopkgtest' entry as an empty
dictionary when it is unable to start the tests.  This triggers us to
believe we have a full test list and stip out all of the MISS entries.
This occurs when we are waiting on a linux-signed package to publish as
we cannot start testing untill all of our dependent set is published.
This leads us to think there are no entries, which can lead us to mark
the testing GOOD until britney gets to the point it can start them.

Detect this present but empty status and ignore it.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>