swm: wft/workflow -- fix accumulation of Holding status
When we accumulate status we are always trying to take the lowest active
phase.  When between phases such as when testing is complete but release
is not yet allowed to start, we enter an inactive phase.  If there is no
later active phase it is interesting but otherwise not.  To handle this we
bump the inactive phases into a high section number.  However, we are using
an inappropriate base value there so that all Holding stati are ignored.

This occured when we renumbered sections to simplify handling.  Fix this
up by introducing a clear definition so we do not do this again.  While
we are here initialise task_text so we always update the task_section
and task_text tuple in lock step simplifying debugging.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>