swm: ignore -- add option to mark invalid
Add a new IgnoreInvalid task target.  This is similar to Ignore except
that as well as ignoring the tasks it also permits move of that task
to Invalid.  This is useful for tasks we have deprecated but we have
existing tracking-bugs carrying those tasks.  With this connected to
the task swm will close them on contact.

Connect this up to the recently deprecated upload-to-ppa-dnu task.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>