cranky: Replace 'source' with '.'
'source' is considered deprecated and its behaviour is shell specific.
'.' is the portable POSIX way of sourcing things.

At the same time silence shellcheck SC1090 and SC1091 (can't follow source).
Where possible use the directive source=./cranky/fixes.d/lib.sh (note that
we use a path relative to the root of the kteam-tools directory, so that
shellcheck can be run from the root dir).

Acked-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>