info/kernel-series: roll bionic/linux-azure to 4.18
Make bionic/linux-azure a backport of cosmic/linux-azure and
bionic/linux-azure-edge a meta only package.

xenial/linux-azure should be kept as a 4.15 kernel, so make it a
regular derivative from bionic/linux. Also fix xenial/linux-azure to
be handled as a meta only package, since 4.18 kernels will not be
brought back to xenial.

Finally, make trusty/linux-azure a backport of xenial/linux-azure
instead of bionic/linux-azure.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>