swm: workflow -- add Ready for phase prefix
We pull out New state as 'Holding before' so we can tell that
a phase is not being actioned because it is not ready.  Once we
are in Confirmed or later we call the phase out as on-going,
but in reality when it is Confirmed it is not actually started.
Call this out as 'Ready for'.

In order to ensure we maintain a determininstic state in the face
of mixtures of New, Confirmed and 'other' statuses we are already
keeping 'Holding before' stati in their own 'between' section range.
Formalise so we can extend this 'between' range to record 'Ready
for' too.  Take this oppotunity to switch to using more obvious
constants for the various ranges.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>