cranky-fix: generate-update-conf: Retrieve DEBIAN_MASTER from kernel-series.yaml
Retrieve DEBIAN_MASTER directly kernel-series.yaml instead of spinning
over the debian.*/ directories. Checking the local directories doesn't
work well, since we might need to rebase the kernel first to get the
updated DEBIAN_MASTER value and we need to update DEBIAN_MASTER in
order to rebase it with cranky-rebase.

We check the parent kernel's name and assume "debian.master" for
"linux" and "debian.BRANCH" for "linux-BRANCH":

Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>