cranky: fix -- provide paths for local master and main package
We have two independent goals when updating packaging fragments within
a tree.  The first is to update components to the official canonical
up-to-date version.  The second is to update components in subsiduary
packages such that they match those in the main package; ie syncing
copies of files in linux-signed to those in linux.

Switch FIX_MASTER to a new cranky/master directory as source for official
canonical versions of package components.  Add a new FIX_MAIN_PATH pointing
to the local checkout of the primary main package (where this package
is a subsiduary package); this will be used to sync package components
between the main package and these subsiduaries.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>