cranky: crl/handle -- add handle names
When reporting errors against handles it is useful to be able to emit
canonical names for handles in play.  Add a new Handle attribute 'name'
which represents a valid handle to instantiate that handle.  This is a
canonical handle and may not be the handle actually used to instantiate it.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Khalid Elmously <khalid.elmously@canonical.com>