cranky: checkout -- add ability to checkout new branches
When we already have a git repository we currently simply add/update the
new remote and tell you the appropriate branch to use; no attempt is made
to check it out.  Add support for checking out the branch.  This tries
hard to ensure we do not lose local uncommitted changes, or reset over
unsaved commits.

Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Khalid Elmously <khalid.elmously@canonical.com>