cranky: master -- getabis fix error handling
When we get a proper network error from wget we erroneously map that to
a group failure, this can lead us to skip the intended grouping and move
on to the next one leaving us with a reduced ABI payload.  Split out
correctly reported failures and network et al failures as intended.

Reported-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Khalid Elmously <khalid.elmously@canonical.com>
Acked-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Signed-off-by: Khalid Elmously <khalid.elmously@canonical.com>