swm: wft/promoter -- handle missing testing tasks
When we are checking if testing is complete (as is done in
promote-to-security) we iterate over the expected testing tasks for this
tracking bug.  If those do not exist we will crater trying to determine
the current status for the absent task.  We do expect these tasks to exist
so if they are missing report this absence as a 'Stalled' status for the
calling task.  The typical caller is using this interface as a silent
'not yet started' marker so this will propogate through to the bug status.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>