ktl/tracking_bug2.py: Fix invalid_tasks handling
Mis-read the meaning of invalid_tasks as defined in kernel_series.yaml
as meaning tasks which should not be created. But in fact it means those
tasks which can be set to invalid.
Luckily the code to do that was still in tasks_reset() for compat
reasons, so all that needs to be done is removing the handling in the
__wf_task_valid() helper.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>