ktl/tracking_bug2.py: Optimize __wf_task_valid() helper
Instead of looking up the source in KernelSeries everytime, do this once
in create(). As series is part of the source object, only the latter
needs to be passed as argument.
Further the launchpadlib series object was only used to decide on the
"upload-to-ppa" task (which no longer exists anyway) and would never
have been unset (because that raises an exception in create()), remove
the case for that and drop another argument.
And since the series object never is unset, some code handling
development series can be simplified, too.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>