swm: wfl/bug -- fix tag matching for main packages
We have an API inconsistency in KernelSeries that the 'main' package is
typically .type==None.  In trying to account for this we are
inadvertantly matching against the empty string and not swm's idea of
'main'.  This could lead us to fail to check for primary tags.  Switch
to using 'main' for swm's state.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>