swm: bug -- only instantiate a master-bug once per run
It is hyper expensive to instantiate a WorkflowBug.  When we reference
the .master_bug property we trigger such an instantiation.  We do this
at least twice for a bug with a master-id set.  Firstly instantiating
this more than once is expensive.  Secondly we now end up with potential
for inconsistent state in each reference which can only lead to issues.

Convert this into a cached property so we only instantiate it once.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>