swm: bug -- expose tracking bug variant
We will soon have multiple variants of tracking bugs; those representing
deb production and those representing snap production.  The variant is
represented via the variant attribute in the swm-properties.  Expose this
via a variant property on the tracking bug object.  Existing bugs without
a variant are marked as 'combo' as they in principle can represent both
at the same time.

Acked-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>