swm: wft/kernel_snap -- disconnect source tracker from snap tracker
Currently we assume that the tracking bug represents both the snap we
are building and the debs it is built from.  Now that we have multiple
snaps associated with a single source we need to disconnect these two.

Split references to attributes of the debs tracker from those to the
snaps tracker using a new .debs_bug for those.  For a legacy 'combo'
variant this simply points to the existing bug.  For a 'snap' variant we
use the master-bug linkage to define the debs tracker from which we will
obtain our binaries; for these we point the .debs_bug at this master bug.

Acked-by: Brad Figg <brad.figg@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>