stable/create-kernel-tasks: Fix task hierarchy
The previous approach turned out to be failing when trying to create
tracking bugs for a re-spin with just 3rd level sources (like hwe-edge
while that is a meta-only derivative of hwe).
In that case the tracking bug was incorrectly linked to the master
instead of hwe as it should.

The re-worked logic will create tasks for every source any requested
source is derived from. Those will be marked as not to be created which
should cause any attempt to spin without existing masters to fail.

While working on this it turned out that things could be a little more
compact when Task().AddSubtask() would return the added task object.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>