cranky-open: Fix exit code when running with --dry-run
Enforce exit code of 0, since the use of the expression:

  [ "$dry_run" -eq 0 ] && ...;

as the last command in the script causes it to finish with exit code
of 1 (error).

Changing the last expression to:

  if [ "$dry_run" -eq 0 ]; then ...; fi

would have the same effect, but adding "exit 0" to the end of the
script can prevent the inclusion of similar errors in the future.

Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>