swm: wft/promote_to_security -- should track promote-to-updates
We are currently duplicating (incompletely) the tests from
promote-to-updates in promote-to-security.  This has recently led a bug to
become promote-to-security in Confirmed when the packages were not even
completly in the proposed pocket.  Basically a tracking bug was reused
without resetting the testing tasks.  Leading to near disaster.

promote-to-security is simply a flag indicating the copy should be to
both the updates and security pockets.  Its criteria should be identicle
to that used in promote-to-updates, with the additional constraint that
it may become separatly invalid if security-signoff is Invalid.

Move to using the state of promote-to-updates as the key for letting
promote-to-security move out of New.  If promote-to-updates is marked
Invalid follow that unconditionally.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>