swm: manager/bug/package -- share KernelSeries instances
It is eminently expensive to instantiate a KernelSeries() object as it
will download the current state of play from launchpad.  It makes more
sense to do this once per run and share that object.  This has the added
advantage that all of the run is guarenteed to have a consistent view of
this data.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>