ktl/sru_cycle: add a SruCycleSpinEntry object
Allow lookup of an SRU cycle using a full spin descriptor.  This returns
a SruCycleSpinEntry object which holds spin and cycle names as well as
any annotations for the cycle from the sru-cycle.yaml.  When a cycle
is not found in the sru-cycle.yaml we still instantiate a result object
but default the values to 'safe' values; for example hold defaults off
if not present in a record, but if there is no record it defaults on.
Also add tests for same.

This is semantically a superset of the existing SruCycleEntry object so
we can kill the existing object and convert the only user.  Update tests
to match.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>