swm: bug -- handle an empty swm-properties section
If the swm-properties section of a tracking bug is empty this will end
up being parsed as None by yaml.safe_load().  This leads us to crater
when examining bug.bprops later on.  We go to a lot of effort elsewhere
in the loader to ensure we always end up with {} at a minimum when there
is no section at all.  Mirror this behaviour when safe_load() returns None.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>