ktl/tracking_bug2.py: Fix unpublished source handling
When a source has not been published to the archive, yet,
creation of a source task fails and we fall back to create
a placeholder task for the "linux" source.
At least that was the initial plan but during on and off
development this was messed up. With this fixed logic it
was possible to create all tracking bugs for bionic which
include 2 source which are not publised, yet.

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>