stable/create-kernel-tasks: Delay setting wf task ready
Right now each tracking bug got the main workflow task set
to ready as soon as it was created. Which is bad as that
signals swm that it can start working on those. But the
create function will first create all derivative and
backport tracking bugs and then register those with the
master bug. And that tries to update swm properties in
the derivative and master tracking bug. Which then has
high chances to fail with precondition errors.
While that should be hardened as well, the quicker solution
is to decouple creation and setting the workflow task to
"in-progress".

Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Andy Whitcroft <andy.whitcroft@canonical.com>
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>