cranky: build-sources -- allow supplying of extra per-package build options
Sometimes it is useful to supply extra, per-package options for
dpkg-buildpackage. For example, when uploading the first kernel
for a new upstream release we need to supply -sa for the kernel
package build but not for other packages.

Add a --build-opts argument to cranky-build-sources to support
this. It takes a string of the format "type:opts", e.g.
"main:-sa", and can be supplied multiple times to supply options
for different package types.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>