create-stable-tracker: strip prefix v on version vX.Y.Z
Most of the time, we will just give a tag instead of a version, like --version
v5.0.2, instead of --version 5.0.2. That same version is used as the bug title,
so some may claim this looks nicer on the changelogs.

However, in order to be more friendly to the user and not fail to grab the logs
because of that, just strip that "v" from the version before adding it again to
run git-log.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>