swm: swm-ls -- expose the target name
When we are reporting on tasks which are about a object other than the
primary source we need to expose the name of that object.  As an example
snap-* tasks represent actions against the snap for a source, for
example a snap-* task on bionic/linux applied to the pc-kernel snap.
swm is now reporting this target so expose it in the output.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>