ktl/kernel_series: supported on sources should inherit from series
As it makes sense to override the series supported status on individual
series we have to examine both always to know the status for a source.
It therefore makes sense that the supported status of a source be
examined relative to the source and only so.  To that end we should
always expose supported on the source and default it to the supported
status of the series.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>