web: gen-workflow-report -- handle snap tracking bugs
Handle the presence of snap tracking bugs as these change the format of
the title some.  Firstly including the series, and secondly including
snap specific data at the end.  Also expose the extra data sections in
the report so you can tell a snap tracker from a debs tracker.  This
also allows you to see which snap a tracker relates to.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>