cranky: master/git-ubuntu-log -- use BugLink trailing text as title
For private kernels some commits are not associated with a public
launchpad bug but rather with a private salesforce ticket. Currently, such
commits are listed as 'miscellaneous' entries in the changelog which is
suboptimal.

To fix that, change the parsing of the BugLink line so that if there's
additional text after the URL, use that as the title for the changelog
entry.

The format is (note the brackets around the title):
  BugLink: <URL> [<TITLE>]

Which will result in the following changelog entry:
  * <TITLE>
    - <COMMIT_SUBJECT>

Example:
  BugLink: https://canonical.my.salesforce.com/1uoSDZ [Fix FOO (SF: #123)]

Which results in:
  * Fix FOO (SF: #123)
    - This commit fixes FOO

While at it, fix flake8 warnings (remove an unused import, add empty lines,
remove unnecessary whitespaces, fix a comment).

Signed-off-by: Juerg Haefliger <juergh@canonical.com>
Acked-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Seth Forshee <seth.forshee@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>