swm: promote-signing-to-proposed -- add a signing step between build and proposed
Add support for a signing PPA sitting between the build-ppa and the
proposed pocket.  This modifies the promote-to-proposed task to request
review and then copying onwards towards proposed.  It then adds a second
task promote-signing-to-proposed to request the second stage copy between
the signing PPA and proposed.

We do this rather than add the more obvious promote-to-signing as this
maintains a simple and obvious workflow task to expected manual handling
mapping for the humans consuming them.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>