swm: package -- new private team compatible source/binary lookup
When determining the current status of a series/package in an archive
pocket we use archive.getBuildSummariesForSourceIds() against the archive.
This interface is intended for use by the archive owner for maintenance.
Sadly this means that permissions checks against require the PPA to be
public, or when private that the caller be a member of the owning team.
It is not sufficient to have an access grant for the PPA.

It seems that you can use getPublishedSources() against the PPA with
only an access grant, and it is possible then to use getBuilds() and
getPublishedBinaries() against those.  Reworked the lookup to use this
vector to confirm the source and binaries are fully published.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>