cranky/docs: cranking-the-kernel refactoring
This is a refactoring in preparation of the new cranking-the-kernel
layout, broken up in four main sections: setup, build, test, review.

Moreover, this new layout will help to automatically generate a
cheatsheet / cookbook processing the same document directly (for example
using a script).

Example of a very trivial one-liner that could be used to generate a
cheatsheet:

 $ grep "^## \|^### " cranky/docs/cranking-the-kernel.md

 ## Setup
 ### Initialize chroot environment - `cranky chroot`
 ## Build
 ### Clone the kernel repository - `cranky clone`
 ### Tool sync stage - `cranky fix`
 ### Rebase stage - `cranky rebase`
 ### Starting commit - `cranky open`
 ### Link to tracker - `cranky link-tb`
 ### Closing commit - `cranky close`
 ### Tagging - `cranky tag`
 ## Test
 ### Testing builds - `cranky test-build`
 ### Prepare meta/signed repositories - `cranky prepare-meta`
 ### Build sources - `cranky build-sources`
 ## Review
 ### Reviewing - `cranky review`
 ### Uploading packages - `cranky upload`

This obviously needs to be improved adding opportune sections in order
to extract the relevant information that we need for each command in
order to produce the cookbook.

Acked-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>
Signed-off-by: Andrea Righi <andrea.righi@canonical.com>
Signed-off-by: Marcelo Henrique Cerri <marcelo.cerri@canonical.com>