swm: bug -- abort correctly on non-workflow bugs or inactive ones
We go to a lot of effort to check that the bug we are cranking is a
workflow bug and marked in-progress, and then we ignore the answers
and just continue to use it.  In the case of a non-workflow bug quickly
exploding for lack of a valid source.

Add appropriate exceptions when the failed states are detected.

Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>