swm: bug/manager -- allow closed bugs to instantiate but not crank
It turns out that the existing prohibitions hinted at by check_is_valid()
are incorrect.  Yes we want to prevent instantiation of non workflow bugs,
but those which are closed can be instantiated if they are master-bugs for
a live bug.  Being closed does not mean they are invalid in the classical
s.is_valid sense, but they shoul never be cranked.  Pull this out to a
new s.is_crankable and use this to prevent accidental scanning.

While we are here, remove the clearly incorrect messaging, and the
now meaningless surrounding if stanza.  Clean up the surrounding loop
short-circuiting it when we see a workflow task; there can be only one.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>