cranky-chroot: install cross libc6-dev packages
Some config options may test and require that the compiler can link to
libc. That requires those packages be installed on the config session
chroot.

In fact, right now, only CONFIG_BPFILTER_UMH requires that, and it requires
not only that cc can link (as the script used for test would indicate), but
that libc is present and can be linked to.

We do not need to add that to Build-Depends, because that is a
build-essential package.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@canonical.com>
Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>