swm: package -- check_component_in_pocket - simplify version checks
We have a number of different package version formats.  Most of them are
distinct and therefore may be checked directly without reference to the
package type.  This helps future proof this support against addition of
new packages.  Note that LBM is a degenerate and ambigious version form
right now, so we must check the type still for this one; note this only
exists in precise/esm.

Acked-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>