cranky: helper-wrapper: fail if no chroot session is found
The main goal of the helper-wrapper script is to find and pass down to
the in-tree scripts the appropriate chroot session. When there's no
session set up, the warning message can be easily ignored and the
scripts could produce an unwanted result by running outside the chroot.

To avoid this issue, always fail when no chroot session is found. Some
of the in-tree scripts don't really use the chroot, but as a chroot is
needed for at least open/close it's safer to always require it. One can
always run the in-tree scripts directly if really wants to run it
outside the chroot.

Acked-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>