cranky: checkout: do not fetch the security repo
Fetching the security repo every time is a waste of time and bandwidth
as the vast majority of cranks are not for a private security issue. The
script also prints every time a warning for the users that don't have
permissions to that repo. And as the script never checks out
automatically any branch from that repo, only fetches it, some manual
work is always needed.

Therefore, it makes more sense to only add/update the security repo to
help in case of a future real need but never fetch it.

Acked-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>