cranky: checkout: use derivative name on local branch
In case the script is checking out a branch that's not from the 'origin'
remote, we use the remote name to compound the local branch name. The
problem with this approach is that sometimes the same derivative repo
can have more than one package (e.g. regular and -edge/-hwe variant) and
using only the remote name doesn't tell which one is checked out.

To make it easier to figure against which package the repo is checked
out, extract the derivative from the package name and use that on the
local branch name instead.

Acked-by: Connor Kuehl <connor.kuehl@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>