cranky: master/git-ubuntu-log: do not duplicate bug entries
If a git commit message references the same bug or CVE more than once we
build a changelog entry with the LP bug title or CVE number repeated as
many times as the number of references. Fix it by converting the initial
list of bugs/CVE's to a set which eliminates the duplicates.

Acked-by: Andy Whitcroft <apw@canonical.com>
Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>