web/Makefile: do not replace populated json report by bare file
When updating the kernel-sru-workflow.json file, it is replaced by the
bare template before the update script is called. During busy times,
the update takes a while and before it finishes other reports that rely
on the same json file (e.g. kernel-stable-board.html) can consume the
bare copy and generate an empty report.

Fix it by working with a tmp file and replacing the previous version of
the populated json only when the new one is finished.

Signed-off-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Connor Kuehl <connor.kuehl@canonical.com>