cranky: Add new script cranky-pull-source
This is a helper script for downloading kernel source packages. For
'regular' kernel packages, it simply calls pull-lp-source under the
covers. For 'special' packages that are not available from the archive
(private or ESM packages), the script will use the URL(s) listed in the
file ~/.getabis.<release>[-<variant>] for the download.

Acked-by: Kleber Sacilotto de Souza <kleber.souza@canonical.com>
Acked-by: Andrea Righi <andrea.righi@canonical.com>
Signed-off-by: Juerg Haefliger <juergh@canonical.com>