swm: wft/promote_to_proposed -- only use signing PPA tasks when we have signed package
When we have a Signing route we unconditionally route promote-to-proposed
copies to Signing, and also liven the promote-signing-to-proposed task
ready for the second stage copy.  This only actually makes sense when the
source has a signed component.  Conditionalise the detection of Signing
on the presence of a signed package, going direct to proposed otherwise.

Acked-by: Stefan Bader <stefan.bader@canonical.com>
Signed-off-by: Andy Whitcroft <apw@canonical.com>